move schedule-related util functions to util/schedule
