+++ site/src/pages/WorkspacePage/WorkspacePage.tsx
@@ -11,7 +11,8 @@ import { FullScreenLoader } from "../../components/Loader/FullScreenLoader"
 import { Workspace, WorkspaceErrors } from "../../components/Workspace/Workspace"
 import { firstOrItem } from "../../util/array"
 import { pageTitle } from "../../util/page"
-import { getFaviconByStatus, maxDeadline, minDeadline } from "../../util/workspace"
+import { maxDeadline, minDeadline } from "../../util/schedule"
+import { getFaviconByStatus } from "../../util/workspace"
 import { selectUser } from "../../xServices/auth/authSelectors"
 import { XServiceContext } from "../../xServices/StateContext"
 import { workspaceMachine } from "../../xServices/workspace/workspaceXService"
